Partition api routes into feature specific files
